requirements.txt
Pillow
pandas
